optimize gcq code
